08 - currency_collector/requirements.txt
beautifulsoup4
lxml
psycopg2
requests
tabula-py
selenium
PyVirtualDisplay
